requirements.txt
fastapi
uvicorn
groq
google-generativeai
PyGithub
unidiff
python-dotenv
pydantic-settings
httpx